Prototype pause() and resume() methods to make the stream processing re-entrant
